youtube_script_writing_tool/requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langchain-google-community
duckduckgo-search
ddgs
streamlit
python-dotenv
